+++ master/buildbot/test/unit/test_buildslave.py
@@ -62,13 +62,14 @@ class AbstractBuildSlave(unittest.TestCase):
             self.ConcreteBuildSlave('bot', 'pass',
                     notify_on_missing=['a@b.com', 13]))
 
-    @defer.deferredGenerator
+    @defer.inlineCallbacks
     def do_test_reconfigService(self, old, old_port, new, new_port):
         master = self.master = fakemaster.make_master()
         old.master = master
-        self.old_registration = old.registration = \
-                mock.Mock(name='old_registration')
-        old.registered_port = old_port
+        if old_port:
+            self.old_registration = old.registration = \
+                    mock.Mock(name='old_registration')
+            old.registered_port = old_port
         old.missing_timer = mock.Mock(name='missing_timer')
         old.startService()
 
@@ -80,12 +81,9 @@ class AbstractBuildSlave(unittest.TestCase):
         new_config.slavePortnum = new_port
         new_config.slaves = [ new ]
 
-        wfd = defer.waitForDeferred(
-            old.reconfigService(new_config))
-        yield wfd
-        wfd.getResult()
+        yield old.reconfigService(new_config)
 
-    @defer.deferredGenerator
+    @defer.inlineCallbacks
     def test_reconfigService_attrs(self):
         old = self.ConcreteBuildSlave('bot', 'pass',
                 max_builds=2,
@@ -102,10 +100,7 @@ class AbstractBuildSlave(unittest.TestCase):
 
         old.updateSlave = mock.Mock(side_effect=lambda : defer.succeed(None))
 
-        wfd = defer.waitForDeferred(
-            self.do_test_reconfigService(old, 'tcp:1234', new, 'tcp:1234'))
-        yield wfd
-        wfd.getResult()
+        yield self.do_test_reconfigService(old, 'tcp:1234', new, 'tcp:1234')
 
         self.assertEqual(old.max_builds, 3)
         self.assertEqual(old.notify_on_missing, ['her@me.com'])
@@ -116,31 +111,53 @@ class AbstractBuildSlave(unittest.TestCase):
         self.assertTrue(old.updateSlave.called)
 
     @defer.deferredGenerator
+    def test_reconfigService_has_properties(self):
+        old = self.ConcreteBuildSlave('bot', 'pass')
+
+        wfd = defer.waitForDeferred(
+            self.do_test_reconfigService(old, 'tcp:1234', old, 'tcp:1234'))
+        yield wfd
+        wfd.getResult()
+
+        self.assertTrue(old.properties.getProperty('slavename'), 'bot')
+
+    @defer.deferredGenerator
+    def test_reconfigService_initial_registration(self):
+        old = self.ConcreteBuildSlave('bot', 'pass')
+
+        wfd = defer.waitForDeferred(
+            self.do_test_reconfigService(old, None, old, 'tcp:1234'))
+        yield wfd
+        wfd.getResult()
+
+        self.assertTrue(self.master.pbmanager.register.called)
+
+    @defer.inlineCallbacks
     def test_reconfigService_reregister_password(self):
         old = self.ConcreteBuildSlave('bot', 'pass')
         new = self.ConcreteBuildSlave('bot', 'newpass')
 
-        wfd = defer.waitForDeferred(
-            self.do_test_reconfigService(old, 'tcp:1234', new, 'tcp:1234'))
-        yield wfd
-        wfd.getResult()
+        yield self.do_test_reconfigService(old, 'tcp:1234', new, 'tcp:1234')
 
         self.assertEqual(old.password, 'newpass')
         self.assertTrue(self.old_registration.unregister.called)
         self.assertTrue(self.master.pbmanager.register.called)
 
-    @defer.deferredGenerator
+    @defer.inlineCallbacks
     def test_reconfigService_reregister_port(self):
         old = self.ConcreteBuildSlave('bot', 'pass')
         new = self.ConcreteBuildSlave('bot', 'pass')
 
-        wfd = defer.waitForDeferred(
-            self.do_test_reconfigService(old, 'tcp:1234', new, 'tcp:5678'))
-        yield wfd
-        wfd.getResult()
+        yield self.do_test_reconfigService(old, 'tcp:1234', new, 'tcp:5678')
 
         self.assertTrue(self.old_registration.unregister.called)
         self.assertTrue(self.master.pbmanager.register.called)
+
+    # FIXME: Test that reconfig properly deals with
+    #   1) locks
+    #   2) telling slave about builder
+    #   3) missing timer
+    # in both the initial config and a reconfiguration.
 
     def test_startMissingTimer_no_parent(self):
         bs = self.ConcreteBuildSlave('bot', 'pass',