+++ master/docs/release-notes.rst
@@ -14,15 +14,20 @@ Master
 Deprecations, Removals, and Non-Compatible Changes
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+* Buildbot master now requires at least Python-2.5 and Twisted-9.0.0.
+
+* Buildbot master requires ``python-dateutil`` version 1.5 to support the
+  Nightly scheduler.
+
 * The configurable callable build.workdir has changed his parameterlist. Instead
   of a single sourcestamp a list of sourcestamps is passed. Each sourcestamp in 
   the list has a different :ref:`codebase<Attr-Codebase>`
 
 * BK support has been removed in this release - see :bb:bug:`2198`.
 
-* The undocumented renderable _ComputeRepositoryURL is no longer imported to
-  py:module::`buildbot.steps.source`. It is still available at
-  py:module::`buildbot.steps.source.oldsource`.
+* The undocumented renderable ``_ComputeRepositoryURL`` is no longer imported to
+  :py:mod:`buildbot.steps.source`. It is still available at
+  :py:mod:`buildbot.steps.source.oldsource`.
 
 Changes for Developers
 ~~~~~~~~~~~~~~~~~~~~~~